+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## [0.5.0-pre]
+- New schema type aliases: int8, uint8, int16, uint16, int32, uint32,
+  int64, uint64, float32, float64.
+
 ## [0.4.3]
 - Fix issue with initbuild.sh for custom builds (#43)
 - Add casts to aid clean C++ builds (#47)